requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator>=2.0
